ITB-48: Appearance fixes. Damaged tests still
