+++ src/components/Order/OrderHistory.css
@@ -1,4 +1,4 @@
 .orderHistory {
+    width: 500px;
     margin: 0 auto;
-    /* padding: 20px 20px 0 20px; */
 }
